backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
channels==4.1.0
channels-redis==4.2.0
python-dotenv==1.0.1
gunicorn==22.0.0
uvicorn==0.29.0